# src/Swagger/TripAppSimple-1.0.0-swagger.yaml
swagger: '2.0'
info:
  description: This is a simple TripApp API
  version: 1.0.0
  title: Simple TripApp API

  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html

# all paths will be relative to the basePath specified
# basePath: /v1


# tags are used for organizing operations
tags:
- name: tickets
  description: Operations for tickets manipulation
- name: validation
  description: Operations for tickets validation  
- name: users
  description: Operations for user management
  
# schemes:
# - http
# - https
paths:
  /tickets:
    get:
      tags: 
        - tickets
      summary: searches tickets purchases
      operationId: searchTickets
      description: |
        By passing in the appropriate options, you can search for
        available ticket in the system
      produces:
      - application/json
      parameters:
      - in: query
        name: searchString
        description: pass an optional search string for looking up tickets
        required: false
        type: string
      - in: query
        name: skip
        description: number of records to skip for pagination
        type: integer
        format: int32
        minimum: 0
      - in: query
        name: limit
        description: maximum number of records to return
        type: integer
        format: int32
        minimum: 0
        maximum: 50
      responses:
        200:
          description: search results matching criteria
          schema:
            type: array
            items:
              $ref: '#/definitions/TicketPurchase'
        400:
          description: bad input parameter
    post:
      tags: 
        - tickets
      summary: adds an ticket purchase item
      operationId: addTicketPurchase
      description: Adds an item to the system
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: ticketPurchase
        description: TicketPurchase item to add
        schema:
          $ref: '#/definitions/TicketPurchase'
      responses:
        201:
          description: item created
        400:
          description: invalid input, object invalid
        409:
          description: item already exists
  /tickets/types:
    get:
      tags: 
        - types
      summary: searches ticket types
      operationId: searchTicketTypes
      description: |
        By passing in the appropriate options, you can search for
        available ticket types in the system
      produces:
      - application/json
      parameters:
      - in: query
        name: searchString
        description: pass an optional search string for looking up ticket types
        required: false
        type: string
      - in: query
        name: skip
        description: number of records to skip for pagination
        type: integer
        format: int32
        minimum: 0
      - in: query
        name: limit
        description: maximum number of records to return
        type: integer
        format: int32
        minimum: 0
        maximum: 50
      responses:
        200:
          description: search results matching criteria
          schema:
            type: array
            items:
              $ref: '#/definitions/TicketType'
        400:
          description: bad input parameter
    post:
      tags: 
        - types
      summary: adds an ticket type
      operationId: addTicketType
      description: Adds an item to the system
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: ticketType
        description: TicketType item to add
        schema:
          $ref: '#/definitions/TicketType'
      responses:
        201:
          description: item created
        400:
          description: invalid input, object invalid
        409:
          description: item already exists
  /tickets/validation:
    get:
      tags: 
        - validation
      summary: searches tickets validations
      operationId: searchValidations
      description: |
        By passing in the appropriate options, you can search for
        available ticket validations in the system
      produces:
      - application/json
      parameters:
      - in: query
        name: searchString
        description: pass an optional search string for looking up ticket validations
        required: false
        type: string
      - in: query
        name: skip
        description: number of records to skip for pagination
        type: integer
        format: int32
        minimum: 0
      - in: query
        name: limit
        description: maximum number of records to return
        type: integer
        format: int32
        minimum: 0
        maximum: 50
      responses:
        200:
          description: search results matching criteria
          schema:
            type: array
            items:
              $ref: '#/definitions/TicketValidation'
        400:
          description: bad input parameter
    post:
      tags: 
        - validation
      summary: adds an ticket validation item
      operationId: addTicketValidation
      description: Adds an item to the system
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: ticketPurchase
        description: TicketValidation item to add
        schema:
          $ref: '#/definitions/TicketValidation'
      responses:
        201:
          description: item created
        400:
          description: invalid input, object invalid
        409:
          description: an existing item already exists
  /codes:
    get:
      tags: 
        - codes
      summary: searches purchase codes
      operationId: searchCodes
      description: |
        By passing in the appropriate options, you can search for
        available purchase codes in the system
      produces:
      - application/json
      parameters:
      - in: query
        name: searchString
        description: pass an optional search string for looking up purchase codes
        required: false
        type: string
      - in: query
        name: skip
        description: number of records to skip for pagination
        type: integer
        format: int32
        minimum: 0
      - in: query
        name: limit
        description: maximum number of records to return
        type: integer
        format: int32
        minimum: 0
        maximum: 50
      responses:
        200:
          description: search results matching criteria
          schema:
            type: array
            items:
              $ref: '#/definitions/PurchaseCode'
        400:
          description: bad input parameter
    post:
      tags: 
        - codes
      summary: adds a purchase code
      operationId: addPurchaseCode
      description: Adds an item to the system
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: purchaseCode
        description: PurchaseCode item to add
        schema:
          $ref: '#/definitions/PurchaseCode'
      responses:
        201:
          description: item created
        400:
          description: invalid input, object invalid
        409:
          description: item already exists
    put:
      tags: 
        - codes
      summary: updates a purchase code
      operationId: updatePurchaseCode
      description: Updates an item to the system
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: purchaseCode
        description: PurchaseCode item to update
        schema:
          $ref: '#/definitions/PurchaseCode'
      responses:
        200:
          description: item updated
        400:
          description: invalid input, object invalid
  /user:
    post:
      tags: 
        - users
      summary: Creates user
      description: This can be done by any user.
      operationId: createUser
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: Created user object
        required: true
        schema:
          $ref: '#/definitions/User'
      responses:
        default:
          description: successful operation
        400:
          description: bad input parameter
        409:
          description: an existing item already exists
  /user/login:
    get:
      tags: 
        - users
      summary: Logs user into the system
      operationId: loginUser
      produces:
      - application/json
      parameters:
      - name: username
        in: query
        description: The user name for login
        required: true
        type: string
      - name: password
        in: query
        description: The password for login in clear text
        required: true
        type: string
      responses:
        200:
          description: successful operation
          schema:
            type: string
          headers:
            X-Rate-Limit:
              type: integer
              format: int32
              description: calls per hour allowed by the user
            X-Expires-After:
              type: string
              format: date-time
              description: date in UTC when token expires
        400:
          description: Invalid username/password supplied
  /user/logout:
    get:
      tags: 
        - users
      summary: Logs out currenty logged in user session
      operationId: logoutUser
      produces:
      - application/json
      parameters: []
      responses:
        default:
          description: successful operation
  /user/{username}:
    get:
      tags: 
        - users
      summary: Gets user by username
      operationId: getUserByUsername
      produces:
      - application/json
      parameters:
      - name: username
        in: path
        description: The username that needs to be fetched.
        required: true
        type: string
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        400:
          description: Invalid username supplied
        404:
          description: User not found
    put:
      tags: 
        - users
      summary: Updates user
      description: This can only be done by the logged in user.
      operationId: updateUser
      produces:
      - application/json
      parameters:
      - name: username
        in: path
        description: username of a user that is about to be updated
        required: true
        type: string
      - in: body
        name: body
        description: Updated user object
        required: true
        schema:
          $ref: '#/definitions/User'
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        400:
          description: Invalid user supplied
        404:
          description: User not found
    delete:
      tags: 
        - users
      summary: Deletes user
      description: This can only be done by an administrator.
      operationId: deleteUser
      produces:
      - application/json
      parameters:
      - name: username
        in: path
        description: The username that needs to be deleted
        required: true
        type: string
      responses:
        200:
          description: successful operation
        400:
          description: Invalid username supplied
        404:
          description: User not found
definitions:
  TicketType:
    required:
    - name
    properties:
      id:
        type: integer
      name:
        type: string
        example: 1 hour ticket
      duration:
        type: integer
        example:  60
        description: duration in minutes
      price:
        type: number
        format: double
        example: 10.50
  TicketPurchase:
    required:
    - id
    - code
    - startDateTime
    - endDateTime
    - type
    - user
    properties:
      id:
        type: integer
      code:
        type: string
        format: uuid
        example: d290f1ee-6c54-4b01-90e6-d701748f0851
      price:
        type: number
        format: double
        example: 50.50
      startDateTime:
        type: string
        format: date-time
        example: 2016-08-29T09:12:33.001Z  
      endDateTime:
        type: string
        format: date-time
        example: 2016-08-29T09:12:33.001Z
      numberOfPassangers:
        type: integer
        description: Number of passangers allowed to travel with buyer and including buyer too.
      type:
        $ref: '#/definitions/TicketType'
      user:
        $ref: '#/definitions/User'
  TicketValidation:
    required:
    - validationDateTime
    - isValid
    - ticket
    - controller
    properties:
      id:
        type: integer
      validationDateTime:
        type: string
        example: 1 hour ticket
      isValid:
        type: boolean
        description: Is ticket valid?
      ticket:
        $ref: '#/definitions/TicketType'
      controller:
        $ref: '#/definitions/User'
  PurchaseCode:
    required:
    - id
    - code
    - value
    - genarationDateTime
    - usageDateTime
    - used
    properties:
      id:
        type: integer
      code:
        type: string
        format: uuid
        example: d290f1ee-6c54-4b01-90e6-d701748f0851
      value:
        type: number
        format: double
        example: 50.50
      genarationDateTime:
        type: string
        format: date-time
        example: 2016-08-29T09:12:33.001Z  
      usageDateTime:
        type: string
        format: date-time
        example: 2016-08-29T09:12:33.001Z
      used:
        type: boolean
        description: Already used by a user.
      user:
        $ref: '#/definitions/User'
  User:
    required:
    - username
    - password
    - role
    properties:
      id:
        type: integer
        format: int64
      username:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      role:
        type: string
      balance:
        type: number
        format: double
        example: 500.00

# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /sergiosuperstar/TripAppSimple/1.0.0
# basePath: /v1/TripAppSimple/1.0.0
schemes:
 - https
 - http